Refactorings: move entrypoint related classes into adapter.entrypoint
