+++ src/main/java/pl/cleankod/exchange/adapter/entrypoint/ExceptionHandlerAdvice.java
@@ -1,10 +1,10 @@
-package pl.cleankod.exchange.entrypoint;
+package pl.cleankod.exchange.adapter.entrypoint;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import pl.cleankod.exchange.core.exception.CurrencyConversionException;
-import pl.cleankod.exchange.entrypoint.model.ApiError;
+import pl.cleankod.exchange.adapter.entrypoint.model.ApiError;
 
 @ControllerAdvice
 public class ExceptionHandlerAdvice {
