+++ src/main/java/pl/cleankod/exchange/adapter/entrypoint/model/ApiError.java
@@ -1,0 +1,4 @@
+package pl.cleankod.exchange.adapter.entrypoint.model;
+
+public record ApiError(String message) {
+}
