--- src/main/java/pl/cleankod/exchange/entrypoint/model/ApiError.java
@@ -1,4 +1,0 @@
-package pl.cleankod.exchange.entrypoint.model;
-
-public record ApiError(String message) {
-}
